+++ ocd/packages/export/src/data/OcdBuildDetails.ts
@@ -11,11 +11,11 @@
 */
 
 export const buildDetails: Record<string, string> = {
-    version: '0.2.9',
-    date: '17/06/2025',
-    time: '16:52:35',
-    datetime: '17/06/2025, 16:52:35',
-    utc: 'Tue, 17 Jun 2025 15:52:35 GMT'
+    version: '0.3.0',
+    date: '26/06/2025',
+    time: '11:31:58',
+    datetime: '26/06/2025, 11:31:58',
+    utc: 'Thu, 26 Jun 2025 10:31:58 GMT'
 }
 
 export default buildDetails